refactor utils into utils.rs
